add version formats to semver
semver seems to only handle versions formatted as 1.2.3 but has issues when versions are formated as 1.02.03 and or if they are windows format with four components. it seems like such version numbers aren't meant to be handled by semver at all but I would still like to propose this change to allow for them.